Feature: Refactor User
