+++ apps/user-service/src/lib/hash.ts
